# templates/golang/{{cookiecutter.project_name}}/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "蓝鲸开发者中心 SaaS 后台 API 文档",
        "title": "BkApp API DOC",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/entries": {
            "get": {
                "tags": [
                    "crud"
                ],
                "summary": "获取条目列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/ginx.PaginatedResp"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "results": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/serializer.EntryListResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "crud"
                ],
                "summary": "创建条目",
                "parameters": [
                    {
                        "description": "创建条目请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.EntryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/entries/{id}": {
            "get": {
                "tags": [
                    "crud"
                ],
                "summary": "获取单个条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serializer.EntryRetrieveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "crud"
                ],
                "summary": "更新条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新条目请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.EntryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "crud"
                ],
                "summary": "删除条目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ginx.PaginatedResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "results": {}
            }
        },
        "ginx.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
...
        "serializer.EntryCreateRequest": {
            "type": "object",
            "required": [
                "categoryID",
                "name",
                "price"
            ],
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "serializer.EntryListResponse": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "serializer.EntryRetrieveResponse": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "serializer.EntryUpdateRequest": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
    }
}